+++ src/components/App/components/public/form/EditableObject/index.js
@@ -37,7 +37,9 @@ export class Editable extends ApolloEditableObject {
       children,
     } = this.props;
 
-    return children ? children.filter(n => n && n.type !== DefaultView) : children;
+    // return children ? children.filter(n => n && n.type !== DefaultView) : children;
+
+    return children;
   }
 
   renderDefaultView() {
@@ -46,7 +48,9 @@ export class Editable extends ApolloEditableObject {
       children,
     } = this.props;
 
-    return children ? children.filter(n => n && n.type !== EditableView) : children;
+    // return children ? children.filter(n => n && n.type !== EditableView) : children;
+
+    return children;
 
   }